move functions out off __init__
Some are moved to backend for lack of a better place for now, most mount-related
functions are moved to a new mount module, though some are left in __init__ for
later, when the dust settles and cyclic imports may become easier to avoid.
